Switch 'Err' to not use continuation-passing style
While the continuation-passing style was sometimes elegant, it was
difficult to understand and extend.  The overhead from wrapping and
unwrapping the replacement 'ErrorT' is probably acceptable, considering
the benefits 'ErrorT' confers.

The functions in 'SAWScript.Compiler' that were being used by other
modules before the switch are still there; their implementations have
changed slightly, but they are backward compatible.  If we want, we can
go think up a better API later on.