Preliminary support for external AIG solvers
An AIG-based solver is expected to return results in the same format as a CNF-based solver. The variable numbers in satisfying assignments should correspond to the inputs to the AIG, counted starting from 1 (since 0 is not a valid CNF variable number).

We may reconsider this and adapt the code to parse the output format given by the AIGER standard, but for the moment a wrapper script will be needed.