Add 'dsec_print' command to SAWScript.
This is done the easy way, by running ABC in a subprocess, not by
adding more bindings to abc-bridge. In particular, the results are
printed and are not available to the SAWScript script. However, this
is way more convenient than choosing file names, using 'write_saig',
and then running 'abc' on the command line yourself.