Make Cryptol expression defaulting less verbose.
The 'SolverConfig' in my last commit is too verbose. This time I
copied a different config from 'Cryptol.ModuleSystem.Env'. I believe
this is the default config that Cryptol actually uses; the last one
was only used for tests. This also adds the '--rewrite-divk' flag, to
support constant division and modulus.