Fix calling convention for uninterpreted functions in SBV parser
SBV files may refer to uninterpreted functions whose return types may
be tuples or records of bitvectors. However, internally to the SBV file,
the structured result must be converted to a single bitvector. It turns
out that the components must be appended so that the left-most bitvector
goes in the *least* significant bit positions.

Previously we had this backwards, which meant that any imported SBV file
with uninterpreted functions had the wrong semantics.