SBVParser: fix combineOutputs handling of n-bit words represented as n 1-bit words
This partially addresses issue #5. Rarely, external SBV files will encode
an output value or argument to an uninterpreted function of type [n] as
a list of n 1-bit words (instead of a single n-bit word, as is common).
Previously, SBVParser would combine the 1-bit words into a list literal,
which was not type correct. This revision instead appends the 1-bit words
to form an n-bit word.

It is still unclear, however, which endianness convention we should use;
I don't have any test cases to tell which way the SBV file expects things
to work. I'm using lsb order for now.