Make test robust to paths containing whitespace.
The value of the right-hand side of a variable assignment is not
subject to word splitting, so we don't need quotes around the outer
command substitution `BASE=$(cygpath ...)`, but we do need to quote
the inner command substitution `$(pwd)` in `...`.

Of course, the current Jenkins workspace path on Windows doesn't have
any spaces in it, so there was not actually a problem.