Port `build-sandbox.sh` to Stack.
Notice how we can run all the tests from the top level package (this
package). I'm not sure if `cabal sandbox` supports this, but it's
nice!

On the other hand, I'm worried that this single `--test-arguments`
option will be trouble if we ever have a path with spaces in it, since
tests indicate that its parts can't be quoted -- rather, we'd have to
escape any embedded spaces.