Implement deep value comparison for Java
This allows Java method specs to be based on deep comparison of the
values reachable from two unequal references. This will be necessary to
allow for specification of methods that allocate new objects. The
objects allocated in the specification and in the implementation  will
likely be different, but should point to the same values.