Rework treatment of assumptions in `java_verify`
The original implementation of the Java simulator tracked two predicates
throughout execution: assertions and assumptions. Assertions were proof
obligations, and needed to be discharged after verification to ensure
the result of simulation was correct. Assumptions were assumed to be
true. When the simulator executed a partial function, it added a
condition to the assertion predicate. When it branched, it added a
condition to the assumption predicate.

Because of this, the original implementation of Java MethodSpecs
interpreted the `java_assert` command (or its equivalent) by adding its
argument to the assertion predicate when executing an override, or to
the assumption predicate when verifying a spec.

Later, we refactored the Java simulator based on the design used by the
LLVM simulator. One aspect of this alternative design was that it no
longer included both an assertion and an assumption predicate. Instead,
only the equivalent of the original assertion predicate existed. Branch
conditions are now tracked implicitly.

In porting the MethodSpecs code to the new Java simulator design, we
conflated assertions and assumptions in a couple of places, leading to
serious soundness issues in the presence of `java_assert`.

This issue likely exists with `llvm_assert`, as well. That will be fixed
in a separate commit.

Fixes #79.