Evaluate Java exprs in local map, not call frame
This changes the functions that read and write Java simulator state at
the locations named by Java expressions to take a local map instead of
an entire call frame as an argument, since that's all it needs. It's
easier to have a local map corresponding to the initial arguments to a
method than it is to have an entire, valid call frame.