Fix a bug in the way LLVM phi functions were being handled.
The semantics of phi functions are a bit subtle, and they were previously
being incorrectly translated.  Basically, a sequence of phi functions at
the beginning of a basic block all need to be evaluated in parallel rather
than one at a time in sequence.  This is important to get the correct values
for loop-carried data dependencies.  Previously "fresh" values would get
propigated into the right-hand-side of phi instructions that occurred textually
later in the program, which gives the incorrect semantics.

The solution is to evaluate the right-hand-side of all phi functions before
doing any assignments.  This ends up giving the correct parallel-move
answers.