Fix broken LLVM struct padding implementation
Now it should work correctly for nested structs. Alignment of an
inner struct is based on the alignment of the whole struct. We don't
treat it as if the fields have been inlined; later fields do not
get placed inside the end-of-struct padding of the previous field.

Also remove the `structPadding` field from StructInfo; store the
end-of-struct padding in the last field of the struct instead.