String together control flow for Go functions in the Crucible translation
There is now simpler handling of return values.  This supports single return
values only, and does not support returning through named return values.  This
change also adds support for conditionals (if statements).  I think the handling
of names in translateBlock is roughly correct...

This change will require a bit of extra handling for dealing with returns
through named return values.  If functions are non-void, but the return value is
missing, we'll have to look up the current values of the return values and
return them at each return site.