Revise the "assignable location" work and implement address-of
The assignable location information is now provided by withTranslatedExpression;
we have to move it here so that the type repr can be carried around with the
translated expression to maintain the equality.  The old formulation had no
connection between them, so the types didn't really work out and required more
existentials.