Finish enough BSV translation to handle a basic AES specification given as a sequences of basic combinatorial circuits.
The translation itself is a bit cobbled together because we take in
the BSV program directly without attempting to do typechecking first.
This makes a number of aspects challenging, including typeclass instance
selection.  In several cases, I resorted to hardcoding the expected output
types in the translation; this happens to work for the AES example in hand.