Change Crucible representation for LLVM function pointers.
They now have the same representation as any other LLVM pointer.
The function handles are stored in the MemImpl using `doMallocHandle`
and are retrieved with a new intrinsic that uses `doLookupHandle`.

The function names also get entries in the `GlobalMap`, so function
symbols can be looked up in the same way as any other global value.