Refactor core crucible datastructures to simplify the API.
ExecutionTree now directly references underlying datastructures instead of
indirecting via a sequence of type families.  This makes the types easier
to navigate and simplifies the module structure somewhat simpler.

In addition, the 'MSSim' module has been deleted, and its contents moved
to other places.  In particular, there is a new 'OverrideSim' module that
contains much of the previous contents of MSSim relating to the override
implementation monad.  Data and utilities related to the global crucible
state have been moved into a separate module named 'GlobalState'.
A small collection of utility datastructures relating to call frames
have been moved into the new 'Frame' module.  Other datastructures have
moved directly into the 'ExecutionTree' module.

In addition, the OverrideSim module, and quite a few other types have
gained a new type parameter 'p', which is threaded through quite a bit
of the system.  This new parameter indicates the 'personality' of the system,
and is intended to be instantiated by the projects that use crucible.
It is a type operator of kind '* -> *' which accepts the 'sym' type, and is
intended to be a packet of state that is threaded thorugh an entire crucible
execution.  It is made avaliable to all override implementations, and allows
users of crucible to maintain their own dynamic state of various sorts.
For example, the MATLAB front end uses its personality state to maintain
data about function load paths and a cache of previously-loaded functions
for dynamic code loading.