Use runtime errors for unsupported LLVM insns
For some types of instructions that would have resulted in a failed
translation to Crucible, we now generate Crucible code that will fail
symbolic simulation if those instructions are executed. This allows us
to load code containing unsupported instructions as long as we don't try
to verify it, or as long as those instructions appear only on infeasible
paths.

This commit doesn't cover every case of failure to translate an
instruction, but does cover the most common cases, and can serve as a
template for any others that come up in practice.