Refactor the handling of merge points in the symbolic execution engine.
This will hopefully fix some persistent bugs relating to the handling of
path conditions.  This merge migrates logic for checking when merges
need to be performed.  Previously, this check was performed when doing
the symbolic branch, and special handling was required for cases where
one of the branch targets was already at the postdominator.  This special-case
code was hard to understand and error-prone.  I am reasonably confident
that it concealed a long-lived bug involving the handling of iterated
merges (positions where multiple symbolic branches need to be resolved at
the same control flow point).

The new code instead delays this check and performs it uniformly whenever
branching to a label.  This reduces the number of special cases and correctly
handles cases where more than one merge is pending at a particular label.