Continue work on compositional verification.
Big update that does parsing and basic processing of a verification
harness.  We parse, rename, typecheck, and transform to SAWCore all
the Cryptol expressions appearing in the harness.  We also reorder
the harness setup steps according to def/use order.

The resulting harness should be in appropriate condition to
allow us to interpret the harness as an override.