Unify the translation-time types of pointers and pointer-width bitvectors.
This allows us to translate programs generated by clang that don't strictly
follow the previous discipline we expected regarding pointers.  E.g., storing
at a pointer type, but later reading it at a bitvector type would cause problems
at translation time.

We now no longer rely on the static type of a value to determine if it is a pointer
or integer.  Instead, that decision is delayed until simulation time.