Change the representation of MIR refefences to better reflect their actual semantics.
One tricky aspect is that MIR is a bit overzealous about allocating
'mutable' local variables.  It seems that any variable that is live
across basic block boundaries is marked mutable; and because
function calls are a block terminator, any value that flows into
the argument of a functionc all becomes mutable.

Instead, we examine the text of a function to determine which
variables have a mutable reference taken.  Only these are
translated as actual referenes; others retain value semantics.