Generalize the LLVM codebase so that pointer widths are no longer a fixed, compile-time value.
All LLVM intrinsics, translation code, etc. that deals with
pointer widths are now abstracted over what the conrecte pointer
size is.  A new constraint `HasPtrWidth` appears in quite a few
type signatures now, and uses implicit parameters to pass around
the runtime representation of the pointer width.

When an LLVM module is translated, its datalayout string is queried
to discover the width of pointers; this information is captured in
the `LLVMContext` record created by `translateModule`.