Simplify readMem so that it doesn't return a separate `Pred sym` value.
Previously, readMem and related functions had a return type of
`(Pred sym, PartLLVMVal sym)`, where the first component is a predicate
indicating whether the result is well-defined, and the second component
is the value read.

However, the `PartLLVMVal` value itself already contains a predicate,
which in practice is always stronger than the `Pred sym` in the first
component (i.e. the `Pred` is always implied by it). In most cases the
two predicates are in fact identical. And then when the result of
`readMem` is used, the two predicates are just and-ed together. Thus
the `Pred` is completely redundant, and has now been removed.