Remove the now-vistigial LLVMMemOps record, and instead just pass around the GlobalVar that contains the LLVM memory.
Also refactor the load and store statements to explicitly take
arguments indicating the Crucible type of the values being
stored and loaded.  Type compatability checking is now done in the
semantics of the statement, rather than outside.  This slightly
reduces the noise of memory loads and stores in the generated
Crucible CFGs.