Fix memory model bug arising when conditional write covers part of a value.
The bug was introduced in revision e242ec3b. The problem was that if you
are reading a large value (e.g. a struct) and a conditional write covers
only part of the range you are reading, then you need to read the *other*
part from the old memory. However the only value we cached was the value
of the *entire struct* read from the old memory. Therefore we end up with
a malformed value that doesn't even have the correct type.

Now, instead of caching a single value, we have a memo table indexed by
type and offset, so we can return the appropriate sub-value. Hopefully
we still avoid the exponential run-time behavior that we had before
e242ec3b.