Major reworking of GEP calculations and constant expression evaluation.
GEP calculations are now processed in two steps.  In the first step
we partially evaluate the GEP and break it down into simpler substeps.
In particular, we figure out any vectorized operations that are occuring
and compute the the number of lanes that are being computed, as well
as the type of the resulting pointer(s).  The resulting GEP intermediate
structure is then evaluated in two different places: one in the computation
of constant expressions, and the other in the ordinary computation of
non-constant GEPs.  Both are made simpler by the introduction of this
new intermedite representation.

As part of this work, the computation of constant expressions has been
significantly improved, made more robust, and vectorized.