crucible-acirc: generateCircuit
The previous version had two major flaws:

  1. It rearranged the input wires (this is really bad!).
  2. If an input to the simulation was a known constant then the
  resulting circuit wouldn't have any corresponding wires. This caused
  issues with using a positional calling convention from Julia.

generateCircuit now takes a 'SimulationResult' that has fields for
the input integers, in order, and also the term that resulted from
simulation. The list of inputs is used to generate circuit input wires
and the terms are used to create the circuit body. This should address
both flaws with the previous version.