crucible-acirc: Remove GroundInt case
By removing this case that we never really used we are able to simplify
circuit construction (see `buildStep` and `doNonceApp`). In particular,
when we see a literal we can now simply create a constant in the circuit
and also have fewer cases to deal with when wiring gates together
(`doNonceApp`).