Continue refactoring/simplifying the handling of configuration options.
There is now an intermediate concept of an `OptionSetting`.  Option settings
can be looked up in a configuration object by name.  Once retrieved, the option
setting object can be used to get or set the current value of the setting.
This avoids the need to look up a setting by name if it will be used frequently,
but also encapuslates the validation and state changes that can occur when
an option is changed.