Reorganize how the interface between the symbolic backend and the symbolic simulator works.
In particular, path conditions are now explicitly managed in a stack.
Context frames are pushed and popped.  Assumptions and assertions
are made in the cotext of a frame; when the frame is popped, all the
associated assumptions are removed.

In addition, we provide a cononical implementation of the assumption stack.
The `SimpleBackend` is then just a straightforward plugging-in of the
assumption stack data structure, and the online backend updates both
the assumption stack, and also mirriors the logical state in the
context of the solver by simultaneously pushing/poping/asserting
formulas.