Move away from naming branches, in favor of using locations.
The names are realliables as sometimes the "true" and "false" in the source
code might correspond to "false" and "true" in the CFG (e.g., if we were
branching on a negated predicate, we could swap the branches around).

Instead we remember the location of the predicate, and the location
of the corresponding block, corresponding to taking the branch.