Stash the value id counter in the CFG
This way `RegRewrite` doesn't need to look through the whole CFG to
figure out where to start generating fresh ids from.

Could also stash the label counter, but that's easier to recalculate and
not needed at the moment. (Also macaw-symbolic uses the CFG constructor
directly and couldn't as easily be adapted.)