Re-re-visit the bug from 52412d2e and fb89069e.
The original bug from 52412d2e was due to a dumb typo.
That code was:

```
mergePartial sym f c (PE px x) (PE py y) =
    do p <- liftIO (itePred sym c px py)
       runPartialT sym p (f p x y)
```

Note that the same predicate `p` is passed both to
`runPartialT` and to `f`.  This is a mistake; `f`
should have been passed the if/then/else condition `c`
instead.

The bugfix from fb89069e took this behavior as intentional and
fixed up one call site in `crucible-llvm` to work around
the bug; however other call sites were not similarly fixed, so
the bug reappeared elsewhere.

This fix essentially reverts fb89069e and fixes the root cause
of the bug.