Add expression parsing library
This will allow much cleaner propagation of relevant error
messages. The current messages are too cryptic, so writing tests is an
exercise in working out why it's all going wrong by reading the
source.

The new library uses the underlying error-location monad from
Culpepper and Felleisen's syntax-parse library, but presents it as a
combinator library rather than a pattern-matching system with
substitutions and bindings.