Significant rearranging of the syntax parser.
The `check` and `synth` phases are now less distinct.  We now
use top-down typing information in the `synth` phase, if it is
avalaible.  In addition, we delay the dispatch of overloaded operators
until such time as type information becomes available (e.g., if
the type of a later argument helps fix the type of earlier arguments).
This significantly reduces the need for type ascriptions, especially
for arithemetic expressions.

Add support for most of the bitvector operations.