Refactor the the way path satisfiability checking works.  We no longer bundle path satisfiablity into the symbolic backends, but instead have separated it into a standalone execution feature.  This allows additional flexability to control how and when path satisfiability is handled.
As part of this refactoring the basic nature of execution features was slightly
refactored.  Instead of being passed continuations, each feature simply
returns a `Maybe` new state.  If `Nothing` is returned, the next feature in the
list is consulted one by one, until we fall into the main simulator loop that advances
the simulator state.  However, if a feature returns a `Just` with a modified state,
then the stack of execution features is restarted from the beginning with the new
state.  This ensures that modified state produced by later execution features in the
list are observed by features earlier in the stack.