Add `Alignment` arguments to `loadRaw` and `loadRawWithCondition`, rather than assuming no alignment.
Refactor `unpackMemValue`, which now takes a `TypeRepr` to fix the
expected output type, rather than packing the result into an `Any`,
and then immediately coercing out. This should improve error messages,
when they occur, and is easier to understand.  The old method was
a vestigial relic of the interface that was necessary before
syntax extensions allowed more precision about how loads would be viewed.