reverseAliases: Fix bug, add unit tests
The bug was that non-alias items weren't added to state.
This would cause all the values in the returned map to be empty.

The unit tests simply exercise a few basic cases. As the comment notes,
it would be nice to have access to Arbitrary instances for the LLVM AST.