Implement a simple way to select between different variants of llvm intrisnics based on their signature.
Essentially, when installing intrinsics that have different signatures
in different LLVM versions, we check if the given signature for the override
appers as a declaration in the LLVM module.  If so, we install the override;
if not we skip it.  We then attempt to install all the known variants of
overrides; at most one should succeed.

The rather naive method we use here requires a linear scan of all declared
functions in the LLVM module, but this probably isn't a problem unless
we encounter modules with a very large number of declared (but not defined)
functions.