Significant refactoring of LLVM intrinsic registration.
LLVM intrinsic registration is now directed by the function
signature declarations that occur in a BC file.  This allows us to
implement the "polymorphic" LLVM intrinsics that are parameterized by
the types of arguments and also to more gracefull handle situations where
the types of LLVM intrisnics differ between versions.

In addition, we now provide implementations for the following new intrinsics:
  - all bitvector variants of `llvm.ctpop`
  - all bitvector variants of `llvm.cttz`
  - all bitvector variants of `llvm.ctlz`
  - all bitvector variants of `llvm.bitreverse`
  - all bitvector variants `llvm.uadd.with.overflow`
  - all bitvector variants `llvm.sadd.with.overflow`
  - all bitvector variants `llvm.usub.with.overflow`
  - all bitvector variants `llvm.ssub.with.overflow`
  - all bitvector variants `llvm.umul.with.overflow`
  - all bitvector variants `llvm.smul.with.overflow`
  - `llvm.bswap.i16`, `llvm.bswap.i32` `llvm.bswap.i48`,
    `llvm.bswap.i64`, `llvm.bswap.i80`, `llvm.bswap.i96`,
    `llvm.bswap.i112`, `llvm.bswap.i128`
  - `llvm.stacksave`, `llvm.stackrestore` (currently no-ops)