Use Nonces in registerized CFG datatypes
Using Ints as the identifiers in atoms, labels, etc. is convenient, but
it makes OrdF and TestEquality instances perform testEquality on
TypeReprs, which is slow enough to get top billing in profiling runs.
(Especially bad with macaw-symbolic, which has tons of values whose type
is a structure with dozens of elements.)  Switching to Nonces gives a
20-25% speedup in Refurbish tests on big PPC binaries.