Return more granular predicates from LLVM mem model operations (#134)
Fixes #133, text copied from that issue:

Currently, memory operations in LLVM.MemModel.Generic return a predicate which
is true if the operation is safe (for some definition of safe). This leads to a
lack of granularity in error messages raised by the callers, for example doFree
may raise the following message:

  let errMsg = "Invalid free (double free or invalid pointer): address " ++ show (G.ppPtr ptr)

The issue is that Generic.freeMem returns an opaque predicate: doFree can't tell
what about this operation actually failed. Additionally, it is easy for the
messages in LLVM.MemModel to inaccurately or incompletely reflect the assertions
made in Generic.