Support varargs SV-COMP functions in crux-llvm
The SV-COMP benchmarks use various verifier control functions that are
given prototypes that sometimes look like `f()` and sometimes like
`f(void)`. The former is considered by Clang to be a varargs function.
We now support all such functions in either form by inspecting the
declared type of the function in the bitcode file and installing the
appropriately-typed override.