Speed up the override registration phase by prefiltering possible overrides first.  This significantly reduces the amount of startup time that is consumed by installing overrides, especially C++ overrides that perform name demangling.
Each override that wants to be registered now provides a "matcher" which examines
the name of a symbol for an approximate test as to weather this override applies.
These can either be exact matches, prefix matches, or substring matches.  Any
override that passes the prefilter test is considered in order, as before, to
determine more precicely if it applies.

On one modestly-sized example, this reduced overall verification time from
105 seconds to 32 seconds.  The algorithm used here for pre-filtering is
quite naive, so additional improvements are perhaps possible.  However,
at the momemnt, startup time is dominated by parsing the bitcode file itself.