what4: Fix SMT generation for z3 in the presence of exists/forall pro… (#235)
* what4: Fix SMT generation for z3 in the presence of exists/forall problems

The previous translation used a more compact notation with multiple values bound
in each let binding:

(let ((a foo)
      (b (bar a))
      ...)

Unfortunately, z3 doesn't scope variables as we might like in this case: the
reference to `a` in the definition of `b` above is not in scope.  This change
uses a single let per value defined instead, which has the expected scoping
behavior.

It seems like this wasn't observed before because the exists/forall backend is
not used very often, and let bindings are only used (by what4) in exists/forall
problems.