Fix Yices hanging when asserting in unsat context (#267)
Yices doesn't like to see `(assert ...)` commands in an unsat context,
and prints a message about this to stderr instead of stdout. Since we're
already tracking whether the context is in an early unsat state, this
commit makes it so we don't emit any more assertions until we've left
the current solver frame.