[crucible-llvm] fix type for atomicrw operation
The atomicrw operation takes a pointer, an op, and a value.  The
pointer must point to a type that is the same as value.  The original
code used the pointer type for the value interpretation, not the
pointed-to type.  This fix uses the pointed-to type (and also provides
an enhanced message for attempts to use an integer value as a
pointer), which fixes errors of the type:

   Branch aborted: AssumedFalse (AssumingNoError Attempted to use integer
   ValInteger NNNN as pointer: i64* in LOC)