WIP commit for adding more realistic symbolic string support.
As I am doing this refactoring, I've (re)discovered that string handling
is stupid and complicated. The avaliable SMT theories for strings treat
strings as a sequence of bytes (as does raw C), but Java wants strings
to be sequences of 16-byte codepoints, and Haskell's `Text` datatype
assumes UTF-8 encoding, etc.

Doing this correctly will requre a lot of care *sigh*.  We'll probably
need to implement a family of string representations to actually
get the semantics right...