Ensure that any necessary type delcarations are carried out before declaring constants/uninterpreted functions/etc that mention them.
Previously, structure types were only declared when a value
of that structure type was created, when meant that sometimes
(e.g., when declaring an uninterpreted function) they would
be mentioned before being declared.

Now, we traverse the structure of types directly before new
constants are declared to ensure that all the necessary types
are declared in scope.