Set the "global declarations" options on CVC4 and Z3 when in online mode, and modify the caching structures in `SMTWriter` so that uninterpreted functions and constants are remembered in all scopes.
This modifies the semantics of push/pop in those solvers so that
declarations are retained across frames.  This allows us to restore
the `DeleteNever` behavior in SMTWriter for top-level constants,
and fixes issue #54.

At some point, we may still wish to revisit the SMTWriter module
to more carefully handle the issue of lexical vs push/pop scoping,
but for now this fixes the soundness issue.