Change the way varargs functions are handled in crucible-llvm.  We used to resolve varagrs calls statically, at call sites, based on the claimed type of the function being called.  However, clang doesn't actually respect the distinction between varargs and non-varargs functions, and assumes that the ABI for function calls is identical for varargs and non-varargs functions depending only on the number and type of actual arguments provided.  In particular, it is happy to bitcast in such a way ignores the distinction, and treats partial external function specifications as though they are varargs.
As a result, we must instead resolve varargs function calls at runtime.  A previous
commit added the minimal necessary support to core crucible.  Here, we ensure that
when we resolve a function pointer into a callable function, we do the necessary
bookeeping so that the variable argument list can be properly packed as expected
by the implementation.