Change `register_llvm_override` so that it allocates the function handle and creates the associated function pointer in the memory at the time the intrinsic is installed.  This in contrast to previous methods, where the handles will be installed a priori, based on the `declare` signature. Now, we instead allocate the handle based on the implementation signature provided with the intrinsic itself.  This matters now that the signatures do not have to match exactly.
To make this work, we had to refactor where `LLVMContext` is defined
so that we can break a module cycle when we reference
`allocLLVMHandleInfo` in `Intrinsics.Common`.

Also, refactor the way that translation and global variable setup works
to remove the antiquated `symbolMap` from the `LLMVContext` datastructure.

The information that used to be contained in that map is now properly stored
in the memory model, and the redundant map was confusing and could be out
of sync with the canonical information in the memory. By finally delaying
the binding of function pointers to function handles, we can correctly
deal with the varargs/partial function signature issue.

Make it easier to get crucible types from an LLVM function declaration.