crux: Allow independent selection of solvers for pathsat checking and goal solving
The behavior when only a single solver is specified remains the same: online
solving with optional support for path satisfiability checking.

This commit adds two new options: --path-sat-solver and
--force-offline-goal-solving.

The --path-sat-solver option lets the user select the solver to use for path
satisfiability checking, which can be different from the solver used to prove
goals.  The --force-offline-goal-solving flag runs the goal solver in
offline mode even if it supports online solving.  If the selected solver does
not support online solving, it will be used in offline mode.

This separation allows us to support solvers without online support (e.g.,
boolector and dreal) for proving goals.  This commit also adds support for the
STP solver as both an online and offline solver.

This commit also adds an option to control the path of the solver interaction
file(s).  It is unfortunate to bundle this up with the larger change.  I could
cherry pick it to a separate commit if we need.  It came up because spawning two
online solver connections contended for the same output file, causing an error.
This commit lets the user specify the interaction files to use.  The default has
changed to no solver interaction files, which avoids littering the filesystem
with slightly mysterious files.