Implement “durable assertions” (#421)
These will not be simplified away and will always show up in the final
list of goals to prove. This commit modifies the overrides for the
various assertion functions supported by `crux-llvm` to use durable
assertions, so that users will always see a goal in the result for each
assertion they explicitly specify in the input source code.