Reorganize the annotation implementation so that the annotation node is a constructor of the `Expr` datatype rather than the `App` datatype. This has a few advantages, among them that the `App` datatype requires fewer type parameters this way.
In addition, we lift the restriction that the annotation type be hashable
and have an equality test. Instead, we allocate a fresh nonce when
`annotateTerm` is called and use that as a proxy for the annotation.
As a result, hash-consing will not coallesce idential terms with
the same annotations.  I expect this will be an uncommon occurence.
However, it means that the equality test for identical terms will
be quick, regardless of the structure of the annotations, and avoids
propigating annoying class restrictions throughout the code.