maintain path info for immutable references
Previously, taking a shared/immutable borrow of a place would construct
a new reference using Const_RefRoot, discarding the place's path
information.  Since the resulting reference has an Empty_RefPath, it
can't be offset, even if it points into an array.  With this commit,
taking an immutable borrow maintains the original root and path of the
place, so offsets are possible.

The downside is that it's no longer possible to mux immutable slices
that point into different arrays.  Previously, this worked because all
immutable references used Const_RefRoot, which can be muxed just like
the underlying value.  Now, they most often use RefCell_RefRoot, which
currently does not support muxing.  If this turns out to be a problem,
then we may need to either support RefCell muxing (which has previously
turned out to be a bit complicated) or find some other workaround (such
as automatically converting to Const_RefRoot when the value is known to
impl Freeze).